+++ hs-abci-test-utils/package.yaml
@@ -20,6 +20,7 @@ description:         Please see the README on GitHub at <https://github.com/gith
 
 dependencies:
 - aeson
+- aeson-pretty
 - base >= 4.7 && < 5
 - bytestring
 - errors
